+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Fixed `SetVerboseDatabaseStatements` losing the script text for `EVAL`/`EVALSHA`
+  commands with StackExchange.Redis 3.2.0+.
+  ([#5223](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5223))
+
 ## 1.18.0-beta.2
 
 Released 2026-Sep-10